[finelog] Drop the store-snapshot transport
`finelog deploy snapshot` tarred a bounded slice of a deployment's store over
`sudo` + `gcloud compute scp` (GCE) or `kubectl exec` (k8s), with catalog
planning and a byte budget behind it. That is a lot of machinery for producing
a directory. `shadow-check` now takes any copy of a store directory, and OPS.md
documents the tar one-liner that makes one.

The archive is not a substitute for the local store dir: it holds compacted
Parquet only, so a `gs://` pull cannot rehearse `.fidx` reads or projection
substitution.
