+++ lib/finelog/AGENTS.md
@@ -189,12 +189,11 @@ would.
 
 `--mode shadow` is how a candidate image is booted against a copy of a real
 store: it serves reads from `--log-dir` and refuses a `gs://`/`s3://` remote or
-a forwarding target at startup. It resolves once into the store's
-`Maintenance`, so no namespace starts a maintenance task — including one
-registered at runtime, which otherwise starts its own. `finelog deploy
-snapshot` and `finelog deploy shadow-check` drive it; see [OPS.md](OPS.md). Use
-it for any local benchmark over a copied store, and pass the mode down rather
-than re-deriving it at a call site.
+a forwarding target at startup. It resolves once into the store's `ServeMode`,
+so no namespace starts a maintenance task — including one registered at
+runtime, which otherwise starts its own. `finelog deploy shadow-check` drives
+it; see [OPS.md](OPS.md). Use it for any local benchmark over a copied store,
+and pass the mode down rather than re-deriving it at a call site.
 
 ## Secondary indexes
